+++ contracts/socket/andromeda-osmosis-token-factory/src/contract.rs
@@ -11,7 +11,6 @@ use andromeda_std::{
     common::{context::ExecuteContext, encode_binary},
     error::ContractError,
 };
-#[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{
     attr, ensure, from_json, to_json_binary, wasm_execute, Addr, Binary, Deps, DepsMut, Env,
@@ -105,19 +104,20 @@ pub fn execute(ctx: ExecuteContext, msg: ExecuteMsg) -> Result<Response, Contrac
             )
         }
         ExecuteMsg::Unlock { recipient } => execute_unlock(ctx, recipient),
-        ExecuteMsg::Receive { msg } => execute_receive(ctx, msg),
+        ExecuteMsg::Receive(msg) => execute_receive(ctx, msg),
         _ => ADOContract::default().execute(ctx, msg),
     }
 }
 
-fn execute_receive(ctx: ExecuteContext, msg: Cw20ReceiveMsg) -> Result<Response, ContractError> {
-    let hook: Cw20HookMsg = from_json(&msg.msg)?;
-
-    match hook {
+fn execute_receive(
+    ctx: ExecuteContext,
+    receive_msg: Cw20ReceiveMsg,
+) -> Result<Response, ContractError> {
+    match from_json(&receive_msg.msg)? {
         Cw20HookMsg::Lock { recipient } => {
             let cw20_addr = ctx.info.sender.clone();
-            let user_addr = ctx.deps.api.addr_validate(&msg.sender)?;
-            let amount = msg.amount;
+            let user_addr = ctx.deps.api.addr_validate(&receive_msg.sender)?;
+            let amount = receive_msg.amount;
             let recipient = recipient
                 .map(|r| r.get_raw_address(&ctx.deps.as_ref()))
                 .transpose()?
@@ -226,7 +226,7 @@ fn execute_unlock(
         };
         SubMsg::new(wasm_execute(
             denom_owner.to_string(),
-            &to_json_binary(&transfer_msg)?,
+            &transfer_msg,
             vec![],
         )?)
     };
@@ -277,7 +277,6 @@ fn execute_create_denom_direct(
     );
 
     DENOMS_TO_OWNER.save(deps.storage, new_denom.clone(), &info.sender)?;
-
     let create_denom_msg = SubMsg::new(MsgCreateDenom {
         sender: env.contract.address.to_string(),
         subdenom: subdenom.clone(),