+++ contracts/socket/andromeda-osmosis-token-factory/src/lib.rs
@@ -1,6 +1,9 @@
 pub mod contract;
 pub mod state;
 
+#[cfg(all(not(target_arch = "wasm32"), feature = "testing"))]
+pub mod mock;
+
 #[cfg(not(target_arch = "wasm32"))]
 mod interface;