+++ tests/e2e/tests/socket_osmo_pool.rs
@@ -26,25 +26,25 @@ fn setup() -> TestCase {
 
     let osmosis_socket_contract = SocketOsmosisContract::new(daemon.clone());
 
-    // // Uncomment this if you want to upload and instantiate a new version of osmosis socket contract
-    // // Make sure to fund the contract after its instantiation
-    // osmosis_socket_contract.upload().unwrap();
-    // osmosis_socket_contract
-    //     .instantiate(
-    //         &andromeda_socket::osmosis::InstantiateMsg {
-    //             kernel_address: "osmo17gxc6ec2cz2h6662tt8wajqaq57kwvdlzl63ceq9keeqm470ywyqrp9qux"
-    //                 .to_string(),
-    //             owner: None,
-    //             swap_router: None,
-    //         },
-    //         None,
-    //         &[],
-    //     )
-    //     .unwrap();
-    // osmosis_socket_contract.set_address(&osmosis_socket_contract.address().unwrap());
-    osmosis_socket_contract.set_address(&Addr::unchecked(
-        "osmo1r2vw2g92f5mt78mj029qlllfsfhrgyh6pzc4zgacllwg7p6x40rqnxgndc".to_string(),
-    ));
+    // Uncomment this if you want to upload and instantiate a new version of osmosis socket contract
+    // Make sure to fund the contract after its instantiation
+    osmosis_socket_contract.upload().unwrap();
+    osmosis_socket_contract
+        .instantiate(
+            &andromeda_socket::osmosis::InstantiateMsg {
+                kernel_address: "osmo17gxc6ec2cz2h6662tt8wajqaq57kwvdlzl63ceq9keeqm470ywyqrp9qux"
+                    .to_string(),
+                owner: None,
+                swap_router: None,
+            },
+            None,
+            &[],
+        )
+        .unwrap();
+    osmosis_socket_contract.set_address(&osmosis_socket_contract.address().unwrap());
+    // osmosis_socket_contract.set_address(&Addr::unchecked(
+    //     "osmo1r2vw2g92f5mt78mj029qlllfsfhrgyh6pzc4zgacllwg7p6x40rqnxgndc".to_string(),
+    // ));
 
     TestCase {
         osmosis_socket_contract,
@@ -65,7 +65,7 @@ fn test_create_pool(setup: TestCase) {
         PoolAsset {
             token: Some(OsmosisCoin {
                 denom: "uosmo".to_string(),
-                amount: "10000000".to_string(),
+                amount: "10000".to_string(),
             }),
             weight: "50000".to_string(),
         },
@@ -92,7 +92,7 @@ fn test_create_pool(setup: TestCase) {
                 pool_params: Some(pool_params),
                 pool_assets,
             },
-            &[coin(1000, "uion"), coin(10000000, "uosmo")],
+            &[coin(1000, "uion"), coin(10000, "uosmo")],
         )
         .unwrap();
     println!("res: {:?}", res);